service/requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-filter==23.1
django-redis==5.2.0
djangorestframework==3.14.0
dnspython==2.3.0
eventlet==0.33.3
flower==1.2.0
gevent==22.10.2
humanize==4.6.0
importlib-metadata==6.1.0
kombu==5.2.4
Markdown==3.4.3
mysqlclient==2.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
tornado==6.2
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
